🚚(frontend) move types and pad api hooks
Move types and pad api hooks to pad-management.
