+++ src/frontend/apps/impress/src/features/pads/pad-management/index.ts
@@ -1,1 +1,3 @@
+export * from './api';
 export * from './components';
+export * from './types';
