+++ src/frontend/apps/impress/src/features/pads/pad-tools/components/ModalPDF.tsx
@@ -11,7 +11,8 @@ import { t } from 'i18next';
 import { useEffect, useState } from 'react';
 
 import { Box, Text } from '@/components';
-import { Pad, usePadStore } from '@/features/pads/pad/';
+import { usePadStore } from '@/features/pads/pad/';
+import { Pad } from '@/features/pads/pad-management';
 
 import { useCreatePdf } from '../api/useCreatePdf';
 import { downloadFile } from '../utils';
