+++ src/frontend/apps/impress/src/features/pads/pad-tools/components/PadToolBox.tsx
@@ -3,8 +3,11 @@ import React, { useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { Box, DropButton, IconOptions, Text } from '@/components';
-import { Pad } from '@/features/pads/pad';
-import { ModalRemovePad, ModalUpdatePad } from '@/features/pads/pad-management';
+import {
+  ModalRemovePad,
+  ModalUpdatePad,
+  Pad,
+} from '@/features/pads/pad-management';
 
 import { TemplatesOrdering, useTemplates } from '../api/useTemplates';
