--- src/frontend/apps/impress/src/features/pads/pad/api/index.ts
@@ -1,1 +1,0 @@
-export * from './usePad';
