+++ src/frontend/apps/impress/src/features/pads/pad/index.tsx
@@ -1,4 +1,2 @@
-export * from './api';
 export * from './components';
 export * from './stores';
-export * from './types';
