--- src/frontend/apps/impress/src/features/pads/pads-panel/api/index.ts
@@ -1,1 +1,0 @@
-export * from './usePads';
