+++ src/frontend/apps/impress/src/features/pads/pads-panel/components/PadItem.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import IconGroup from '@/assets/icons/icon-group.svg';
 import { Box, StyledLink, Text } from '@/components';
 import { useCunninghamTheme } from '@/cunningham';
-import { Pad } from '@/features/pads/pad';
+import { Pad } from '@/features/pads/pad-management';
 
 import IconNone from '../assets/icon-none.svg';
