+++ src/frontend/apps/impress/src/features/pads/pads-panel/index.ts
@@ -1,2 +1,1 @@
-export * from './api';
 export * from './components';
