+++ src/frontend/apps/impress/src/pages/pads/[id].tsx
@@ -5,7 +5,8 @@ import { ReactElement } from 'react';
 
 import { Box } from '@/components';
 import { TextErrors } from '@/components/TextErrors';
-import { PadEditor, usePad } from '@/features/pads/pad';
+import { PadEditor } from '@/features/pads/pad';
+import { usePad } from '@/features/pads/pad-management';
 import { PadLayout } from '@/layouts';
 import { NextPageWithLayout } from '@/types/next';
